[minor] move `Browser` DOM/interactive methods to extensions
