+++ src/Browser/Extension/DomExtension.php
@@ -1,0 +1,99 @@
+<?php
+
+namespace Zenstruck\Browser\Extension;
+
+/**
+ * @author Kevin Bond <kevinbond@gmail.com>
+ */
+trait DomExtension
+{
+    /**
+     * @return static
+     */
+    final public function assertSee(string $expected): self
+    {
+        return $this->wrapMinkExpectation(
+            fn() => $this->webAssert()->pageTextContains($expected)
+        );
+    }
+
+    /**
+     * @return static
+     */
+    final public function assertNotSee(string $expected): self
+    {
+        return $this->wrapMinkExpectation(
+            fn() => $this->webAssert()->pageTextNotContains($expected)
+        );
+    }
+
+    /**
+     * @return static
+     */
+    final public function assertSeeIn(string $selector, string $expected): self
+    {
+        return $this->wrapMinkExpectation(
+            fn() => $this->webAssert()->elementTextContains('css', $selector, $expected)
+        );
+    }
+
+    /**
+     * @return static
+     */
+    final public function assertNotSeeIn(string $selector, string $expected): self
+    {
+        return $this->wrapMinkExpectation(
+            fn() => $this->webAssert()->elementTextNotContains('css', $selector, $expected)
+        );
+    }
+
+    /**
+     * @return static
+     */
+    final public function assertSeeElement(string $selector): self
+    {
+        return $this->wrapMinkExpectation(
+            fn() => $this->webAssert()->elementExists('css', $selector)
+        );
+    }
+
+    /**
+     * @return static
+     */
+    final public function assertNotSeeElement(string $selector): self
+    {
+        return $this->wrapMinkExpectation(
+            fn() => $this->webAssert()->elementNotExists('css', $selector)
+        );
+    }
+
+    /**
+     * @return static
+     */
+    final public function assertElementCount(string $selector, int $count): self
+    {
+        return $this->wrapMinkExpectation(
+            fn() => $this->webAssert()->elementsCount('css', $selector, $count)
+        );
+    }
+
+    /**
+     * @return static
+     */
+    final public function assertElementAttributeContains(string $selector, string $attribute, string $expected): self
+    {
+        return $this->wrapMinkExpectation(
+            fn() => $this->webAssert()->elementAttributeContains('css', $selector, $attribute, $expected)
+        );
+    }
+
+    /**
+     * @return static
+     */
+    final public function assertElementAttributeNotContains(string $selector, string $attribute, string $expected): self
+    {
+        return $this->wrapMinkExpectation(
+            fn() => $this->webAssert()->elementAttributeNotContains('css', $selector, $attribute, $expected)
+        );
+    }
+}
