+++ src/Browser/PantherBrowser.php
@@ -8,6 +8,7 @@ use Symfony\Component\Panther\Client;
 use Symfony\Component\VarDumper\VarDumper;
 use Zenstruck\Assert;
 use Zenstruck\Browser;
+use Zenstruck\Browser\Extension\InteractiveExtension;
 use Zenstruck\Browser\Mink\PantherDriver;
 use Zenstruck\Browser\Response\PantherResponse;
 use Zenstruck\Callback\Parameter;
@@ -19,6 +20,8 @@ use Zenstruck\Callback\Parameter;
  */
 class PantherBrowser extends Browser
 {
+    use InteractiveExtension;
+
     private Client $client;
     private ?string $screenshotDir = null;
     private ?string $consoleLogDir = null;
